# Implement dedup-ready label sorting in stores