--- a/cmd/thanos/rule.go
+++ b/cmd/thanos/rule.go
@@ -607,10 +607,11 @@ func runRule(
 				if httpProbe.IsReady() {
 					mint, maxt := tsdbStore.TimeRange()
 					return &infopb.StoreInfo{
-						MinTime:           mint,
-						MaxTime:           maxt,
-						SupportsSharding:  true,
-						SendsSortedSeries: true,
+						MinTime:                        mint,
+						MaxTime:                        maxt,
+						SupportsSharding:               true,
+						SendsSortedSeries:              true,
+						SendsSortedSeriesWithoutLabels: true,
 					}
 				}
 				return nil