--- a/cmd/thanos/store.go
+++ b/cmd/thanos/store.go
@@ -397,10 +397,11 @@ func runStore(
 			if httpProbe.IsReady() {
 				mint, maxt := bs.TimeRange()
 				return &infopb.StoreInfo{
-					MinTime:           mint,
-					MaxTime:           maxt,
-					SupportsSharding:  true,
-					SendsSortedSeries: true,
+					MinTime:                        mint,
+					MaxTime:                        maxt,
+					SupportsSharding:               true,
+					SendsSortedSeries:              true,
+					SendsSortedSeriesWithoutLabels: true,
 				}
 			}
 			return nil