--- a/pkg/info/infopb/rpc.proto
+++ b/pkg/info/infopb/rpc.proto
@@ -56,8 +56,12 @@ message StoreInfo {
     int64 min_time = 1;
     int64 max_time = 2;
     bool  supports_sharding = 3;
+    // TODO(fpetkovski): Remove in v1.0
     bool  sends_sorted_series = 4;
+    // TODO(fpetkovski): Remove in v1.0
+    bool  sends_sorted_series_without_labels = 5;
 }
+
 
 // RulesInfo holds the metadata related to Rules API exposed by the component.
 message RulesInfo {