--- a/pkg/query/test.go
+++ b/pkg/query/test.go
@@ -657,6 +657,7 @@ func (i inProcessClient) SupportsSharding() bool {
 	return false
 }
 
-func (i inProcessClient) SendsSortedSeries() bool { return false }
-func (i inProcessClient) String() string          { return i.name }
-func (i inProcessClient) Addr() (string, bool)    { return i.name, true }
+func (i inProcessClient) SendsSortedSeries() bool         { return false }
+func (i inProcessClient) String() string                  { return i.name }
+func (i inProcessClient) Addr() (string, bool)            { return i.name, true }
+func (i inProcessClient) SendsSeriesSortedForDedup() bool { return false }