This commit adds functionality to all stores to sort series labels in order appropriate for deduplication. This is done by propagating replica labels from the Querier in the Series request. When all stores send a hint that they support sorting labels, we can rely on the k-way merge in the proxy and avoid an expensive global sort.

For backwards compatibility, each Store will send a hint in the SeriesResponse to indicate that it supports this functionality. The hint is sent at the beginning of the stream, and is the first response that each store will send.

The Store Proxy will inspect the hints from each store and detect cases where a store does not support resorting of labels. In that case it will itself send a hint upstream that sorting is not supported.

Fixes https://github.com/thanos-io/thanos/issues/5719
Closes https://github.com/thanos-io/thanos/pull/5742 https://github.com/thanos-io/thanos/pull/5692

<!--
    Keep PR title verbose enough and add prefix telling
    about what components it touches e.g "query:" or ".*:"
-->

<!--
    Don't forget about CHANGELOG!

    Changelog entry format:
    - [#<PR-id>](<PR-URL>) Thanos <Component> ...

    <PR-id> Id of your pull request.
    <PR-URL> URL of your PR such as https://github.com/thanos-io/thanos/pull/<PR-id>
    <Component> Component affected by your changes such as Query, Store, Receive.
-->

* [ ] I added CHANGELOG entry for this change.
* [x] Change is not relevant to the end user.

## Changes

<!-- Enumerate changes you made -->

## Verification

<!-- How you tested it? How do you know it works? -->