@bwplotka 2022-10-18
Actually I like @GiedriusS idea - we really need to tell WHAT gRPC version it is (if new aspect of GRPC feature is supported)... Adding bool to info is enough indeed, but maybe simply add version of gRPC to info? Something naive like that would help us to determine the feature set of store API 🤔