@bwplotka 2022-10-18
> The problem is that we cannot know in advance if a store is able to send data sorted for dedup. Since we stream series one by one from store/tsdb.go and store/prometheus.go (receiver, ruler and sidecar), data from these stores will be sorted according to the TSDB specific sort (according to labels.Compare).
> Sorting for dedup depends on replica labels, and if a user sets replica labels that are regular series labels in TSDB, these stores will not be able to send series in order required for dedup. So whether a store has this capability or not depends on which labels are set as replica labels.

I don't think we have this problem. We want to add guarantee of Store API that if certain field is set data is sorted differently. So we need only simple information - does certain Store API supports this new guarantee or not?