@fpetkovski 2022-10-19
@GiedriusS we spoke with @bwplotka in Slack about adding buffering to the sidecar and receiver when replica labels are not external labels, which is the second solution from this issue: https://github.com/thanos-io/thanos/issues/5719. 

In this case we will load all series in memory in these components, sort them without taking replica labels into account, and only then send upstream. This will allow us to use the static checks from your suggestion and avoid the complicated response hints. 

When replica labels are external labels (which is the most common use case), we won't need to buffer and we can stream everything as we do now. Wdyt?