@fpetkovski 2022-10-19
I pushed a new commit that implements this idea. We can add a [server proxy](https://github.com/thanos-io/thanos/pull/5796/files#diff-2071cab2d2833276f565f921c5123141a9a46f6a07818c9ca26f2f83d54ba7b1) that can decide whether to bufffer and re-sort or not. This way we keep all stores as they are and avoid various checks that can be error prone and brittle.