@fpetkovski 2022-10-20
Here are some benchmarks

Select with 2 replicas
```
name                                             old time/op    new time/op    delta
QuerySelect/1000000SeriesWith1Samples/select-8      1.10s ± 6%     0.84s ± 2%  -23.52%  (p=0.008 n=5+5)
QuerySelect/100000SeriesWith100Samples/select-8     110ms ± 1%      96ms ± 3%  -12.78%  (p=0.008 n=5+5)
QuerySelect/1SeriesWith10000000Samples/select-8    19.5ms ± 1%    20.3ms ± 1%   +3.99%  (p=0.008 n=5+5)

name                                             old alloc/op   new alloc/op   delta
QuerySelect/1000000SeriesWith1Samples/select-8      818MB ± 0%     743MB ± 0%   -9.20%  (p=0.008 n=5+5)
QuerySelect/100000SeriesWith100Samples/select-8    82.5MB ± 0%    74.4MB ± 0%   -9.88%  (p=0.008 n=5+5)
QuerySelect/1SeriesWith10000000Samples/select-8    18.7MB ± 0%    18.7MB ± 0%   +0.06%  (p=0.008 n=5+5)

name                                             old allocs/op  new allocs/op  delta
QuerySelect/1000000SeriesWith1Samples/select-8      12.5M ± 0%      9.6M ± 0%  -23.15%  (p=0.008 n=5+5)
QuerySelect/100000SeriesWith100Samples/select-8     1.25M ± 0%     0.96M ± 0%  -23.32%  (p=0.008 n=5+5)
QuerySelect/1SeriesWith10000000Samples/select-8      167k ± 0%      167k ± 0%   +0.11%  (p=0.008 n=5+5)
```

Select with 5 replicas
```

name                                             old time/op    new time/op    delta
QuerySelect/1000000SeriesWith1Samples/select-8      1.04s ± 6%     0.86s ± 3%  -16.72%  (p=0.008 n=5+5)
QuerySelect/100000SeriesWith100Samples/select-8     105ms ± 2%      97ms ± 1%   -7.52%  (p=0.008 n=5+5)
QuerySelect/1SeriesWith10000000Samples/select-8    19.1ms ± 1%    19.6ms ± 1%   +2.52%  (p=0.008 n=5+5)

name                                             old alloc/op   new alloc/op   delta
QuerySelect/1000000SeriesWith1Samples/select-8      814MB ± 0%     752MB ± 0%   -7.54%  (p=0.008 n=5+5)
QuerySelect/100000SeriesWith100Samples/select-8    82.0MB ± 0%    75.0MB ± 0%   -8.60%  (p=0.008 n=5+5)
QuerySelect/1SeriesWith10000000Samples/select-8    18.7MB ± 0%    18.7MB ± 0%   +0.10%  (p=0.008 n=5+5)

name                                             old allocs/op  new allocs/op  delta
QuerySelect/1000000SeriesWith1Samples/select-8      12.2M ± 0%      9.2M ± 0%  -24.59%  (p=0.008 n=5+5)
QuerySelect/100000SeriesWith100Samples/select-8     1.22M ± 0%     0.92M ± 0%  -24.55%  (p=0.008 n=5+5)
QuerySelect/1SeriesWith10000000Samples/select-8      167k ± 0%      167k ± 0%   +0.18%  (p=0.016 n=5+4)

```