@fpetkovski 2022-10-24
One thing I realized is that the whole complexity comes from the fact that users are allowed to dedup by a label that is in TSDB and is not necessarily an external label. I wonder if this is even a valid use case, and how many people need to do something like this. Dropping support for this might break some users, but maybe we can add a flag in the querier `query.dedup-label-in-tsdb` or something similar which can be propagated to stores to tell them to resort the data.