@bwplotka 2022-10-25
So first of all, we have a bug in the PromQL engine, so had to ensure this PR is correct: https://github.com/thanos-io/thanos/issues/5820

Second is - I think I would be happy to fail or not sort for store API implementation that "label" to sort without is not an external label but actually inside e.g. block. so - for this example:

![image](https://user-images.githubusercontent.com/6950331/197884771-1b256c95-0ed8-4cd8-9e0e-82ff65198fa0.png)

But in either case we have to check actively sorting, so:

1. If we want to fail, sidecar/receive/store etc has to actively check if label passed in request is part of replica label or it exists in block etc.
2. If we want to not sort, querier has to do check if receive sorted stream.

Also  - if we don't pass any label and there ARE replica labels for e.g. prometheus+sidecar or receive etc - we return unsorted series, right? so if we don't want to buffer/sort on leafs - we can't really guarantee sorted results on Store API? -- so, is you PR works with no replica labels passed on querier?

Anyway, sorry for back-forth, but it's a big thing spanning multiple components and it's very easy to break other people setups.