@bwplotka 2022-10-25
Now I am leaning more into hinting sorting and checking this on querier - sounds like easier way... 🤔 

I am preparing for demo, but will look at this after Wed.