@GiedriusS 2023-02-20
I think we can close this since an alternative implementation was merged for solving this problem? :smile: